+++ src/lib.rs
@@ -60,6 +60,11 @@ pub mod sdk_source;
 #[cfg(feature = "native-sdk")]
 #[cfg_attr(docsrs, doc(cfg(feature = "native-sdk")))]
 #[cfg(any(target_os = "windows", target_os = "linux"))]
+pub mod sdk_sink;
+
+#[cfg(feature = "native-sdk")]
+#[cfg_attr(docsrs, doc(cfg(feature = "native-sdk")))]
+#[cfg(any(target_os = "windows", target_os = "linux"))]
 pub use native_sdk::{NativeSdkClient, NativeSdkSource};
 
 #[cfg(feature = "ffi")]